+++ src/__tests__/setup.ts
@@ -1,7 +1,7 @@
-import mockLogger from "./mockLogger";
+import { logger } from '../__mocks__/logger';
 
 // Mock the Pino logger
-jest.mock('pino', () => () => mockLogger);
+jest.mock('pino', () => () => logger);
 
 // Always mock out our config files
 jest.mock('../config/generalConfig', () => ({
@@ -10,23 +10,3 @@ jest.mock('../config/generalConfig', () => ({
     jwtTtl: 5,
   }
 }));
-
-jest.mock('../config/oauthConfig', () => ({
-  oauthConfig: {
-    authorizationCodeLifetime: 10,
-    accessTokenLifetime: 30,
-    refreshTokenLifetime: 30,
-  }
-}))
-
-// Mock the MySQL connection
-jest.mock('../datasources/mySQLDataSource', () => {
-  return {
-    __esModule: true,
-    MySQLDataSource: {
-      getInstance: jest.fn().mockReturnValue({
-        query: jest.fn(),
-      }),
-    },
-  };
-});